+++ src/main/java/kr/hhplus/be/server/interfaces/coupon/CouponControllerDocs.java
@@ -9,10 +9,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kr.hhplus.be.server.domain.user.User;
-import kr.hhplus.be.server.support.response.ResultResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.ErrorResponse;
-import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
